Add 2 Files
debuggin-with-git.txt

submodules.txt

Debuggin With Git was neat and readable.

But Submodules Had a ton of info. I wonder if I should stop taking notes and just read through the rest of the rest of the book. In a real project commit, I would write things like this, but this is just practice.