Add 2 Files
Both files do not have any notes for the Pro Git book chapter 9 because I decided not to read them since I don't need to read that information at the moment, but if I ever use Git as a client or have to migrate a project that requires me to migrate it to Git, I'll read it.

Though I didn't take notes in those files, I used the two files as an opportunity to practice using "git add -i" for interactively adding files and such.