Add .gitignore file
The .gitignore file was created at the beginning, but wasn't added because it wasn't really needed since I only use to prevent Git from automatically adding and showing untracked files on my local, remote computer.

I'm adding it now because it I believe it belongs with the rest of the files I created for practicing the main concepts in Git to help enable me to do remote work with someone, a team, or company.